Settings: Abandon the App Ops Template structure
We aren't really using the OpsTemplate anymore. Previously it was a
wrapper made to hold two constant arrays. As we are just using one
of these templates, we might as well switch to simple arrays instead.
This means we don't have to pass around any objects anywhere or worry
about which one we should use.

tl;dr Follow-up for single page merge commit.

Change-Id: I984333e60a3ab7ef31c5ea63e1c8a396bec4093a